requirements.txt
appdirs==1.4.3
astroid==2.3.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
google-cloud-ndb==1.2.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.17
Werkzeug==1.0.1
wrapt==1.11.2